requirements.txt
amqp==5.1.1
appdirs==1.4.3
asgiref==3.5.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.2.6
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==4.0.4
django-configurations==2.3.2
django-constance==2.9.0
django-cors-headers==3.12.0
django-filter==21.1
django-guardian==2.4.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
fancycompleter==0.9.1
flake8==4.0.1
flower==1.0.0
gunicorn==20.1.0
html5lib==1.0.1
humanize==4.1.0
idna==2.8
ipaddr==2.2.0
kombu==5.2.4
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pdbpp==0.10.3
pep517==0.8.2
progress==1.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==2.4.6
pyrepl==0.9.0
pytoml==0.1.21
pytz==2022.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wmctrl==0.4
